Reorganize API.
- break up into files
- do not use custom `struct` for location-value pairs, just specify
  the API
- allow closures as local methods
- load NLopt code on demand
